finish minimal inference tab; make replay able to take pickle in addition to dats for faster load time
