+++ Presets/LSLPresets/PredictionResults.json
@@ -1,0 +1,5 @@
+{ "StreamName":"PredictionResult",
+  "ChannelNames":[
+    "P(Distractor)", "P(Target)",
+    "P(Novelty)"],
+  "NominalSamplingRate":  50}
